Refactor table reading and simplify image verification (#57)
Remove the caching mechanism that 1) didn't work properly in the first
place and 2) could technically be incorrect if images have aliases and
are aliased to a different copy of the images.

Move the call to `verify_image` to the already existing loop that reads
data from the 3LC Table, and make a common loop for classification and
detection dataset classes.

Ignore bounding boxes with nonpositive width or height, akin to what is
already done for instance segmentations with nonpositive area.