Allow tables without a weight column to be used unless sample weighting is enabled for training (#63)
Remove usage of `tlc.SAMPLE_WEIGHT` as the weight column name, instead infer it with `table.weights_column_name`.

If exclusion of zero weight samples is enabled for training or collection, but no weight column exists, no samples are excluded and the run proceeds. Only when `sampling_weights=True` a `ValueError` is raised.